+++ ui-v2/app/templates/components/consul-service-list.hbs
@@ -1,0 +1,37 @@
+{{#if (gt items.length 0)}}
+  {{#tabular-collection items=items as |item index|}}
+    {{#block-slot name='header'}}
+      <th style={{remainingWidth}}>Service</th>
+      <th style={{totalWidth}}>
+        Health Checks
+        <span>
+          <em role="tooltip">The number of health checks for the service on all nodes</em>
+        </span>
+      </th>
+      <th style={{remainingWidth}}>Tags</th>
+    {{/block-slot}}
+    {{#block-slot name='row'}}
+      <td data-test-service={{item.Name}} style={{remainingWidth}}>
+        <a href={{href-to routeName item.Name}}>
+      {{#let (service/external-source item) as |externalSource| }}
+        {{#if externalSource }}
+          <span data-test-external-source={{externalSource}} style={{concat 'background-image: var(--' externalSource '-icon)'}}></span>
+        {{else}}
+          <span></span>
+        {{/if}}
+      {{/let}}
+          {{item.Name}}
+        </a>
+      </td>
+      <td style={{totalWidth}}>
+        {{healthcheck-info
+          passing=item.ChecksPassing warning=item.ChecksWarning critical=item.ChecksCritical
+          passingWidth=passingWidth warningWidth=warningWidth criticalWidth=criticalWidth
+        }}
+      </td>
+      <td style={{remainingWidth}}>
+        {{tag-list items=item.Tags}}
+      </td>
+    {{/block-slot}}
+  {{/tabular-collection}}
+{{/if}}
